Move strip_session_routing_state into auth module
Colocate session file manipulation with the rest of the session
persistence logic. Uses atomic_write (temp + rename) instead of bare
tokio::fs::write, uses the canonical HashMap<String, String> type,
and handles serialisation errors instead of silently swallowing them.
