+++ src/auth/session.rs
@@ -125,6 +125,49 @@ async fn atomic_write(path: &Path, data: &[u8]) -> Result<()> {
     Ok(())
 }
 
+/// Strip routing state from a persisted session file, keeping only `trust_token`
+/// and `client_id`.
+///
+/// Clearing `session_token` forces `authenticate()` through the SRP path instead
+/// of the validate shortcut. Preserving `trust_token` lets SRP send it in
+/// `trustTokens`, so Apple can recognise a trusted device and skip 2FA.
+///
+/// Deletes the file if it is corrupt or unreadable (falling back to the old
+/// delete-everything behaviour).
+pub(crate) async fn strip_session_routing_state(session_file: &Path) {
+    let contents = match fs::read_to_string(session_file).await {
+        Ok(c) => c,
+        Err(e) if e.kind() == std::io::ErrorKind::NotFound => return,
+        Err(_) => {
+            let _ = fs::remove_file(session_file).await;
+            return;
+        }
+    };
+
+    let mut map: HashMap<String, String> = match serde_json::from_str(&contents) {
+        Ok(m) => m,
+        Err(_) => {
+            let _ = fs::remove_file(session_file).await;
+            return;
+        }
+    };
+
+    map.retain(|k, _| k == "trust_token" || k == "client_id");
+
+    match serde_json::to_string_pretty(&map) {
+        Ok(json) => {
+            if let Err(e) = atomic_write(session_file, json.as_bytes()).await {
+                tracing::warn!(error = %e, "Could not rewrite session file, removing");
+                let _ = fs::remove_file(session_file).await;
+            }
+        }
+        Err(e) => {
+            tracing::warn!(error = %e, "Could not serialise stripped session, removing");
+            let _ = fs::remove_file(session_file).await;
+        }
+    }
+}
+
 /// Build the API and download HTTP clients for a session.
 ///
 /// Both share the same cookie jar. The API client uses a total-request timeout;
@@ -1040,5 +1083,70 @@ mod tests {
             Duration::from_secs(45),
             "reset_http_clients must preserve the configured timeout"
         );
+    }
+
+    #[tokio::test]
+    async fn strip_session_preserves_trust_token_and_client_id() {
+        let dir = tempfile::tempdir().unwrap();
+        let path = dir.path().join("test.session");
+        let data = serde_json::json!({
+            "session_token": "tok_abc",
+            "session_id": "sid_123",
+            "scnt": "scnt_val",
+            "trust_token": "trust_xyz",
+            "client_id": "auth-1234",
+            "account_country": "USA"
+        });
+        std::fs::write(&path, data.to_string()).unwrap();
+
+        strip_session_routing_state(&path).await;
+
+        let contents = std::fs::read_to_string(&path).unwrap();
+        let map: HashMap<String, String> = serde_json::from_str(&contents).unwrap();
+        assert_eq!(map.len(), 2);
+        assert_eq!(map["trust_token"], "trust_xyz");
+        assert_eq!(map["client_id"], "auth-1234");
+        assert!(!map.contains_key("session_token"));
+        assert!(!map.contains_key("session_id"));
+        assert!(!map.contains_key("scnt"));
+    }
+
+    #[tokio::test]
+    async fn strip_session_corrupt_file_gets_deleted() {
+        let dir = tempfile::tempdir().unwrap();
+        let path = dir.path().join("corrupt.session");
+        std::fs::write(&path, "not valid json {{{").unwrap();
+
+        strip_session_routing_state(&path).await;
+
+        assert!(!path.exists());
+    }
+
+    #[tokio::test]
+    async fn strip_session_missing_file_is_noop() {
+        let dir = tempfile::tempdir().unwrap();
+        let path = dir.path().join("nonexistent.session");
+
+        strip_session_routing_state(&path).await;
+
+        assert!(!path.exists());
+    }
+
+    #[tokio::test]
+    async fn strip_session_no_trust_token_leaves_empty_file() {
+        let dir = tempfile::tempdir().unwrap();
+        let path = dir.path().join("no_trust.session");
+        let data = serde_json::json!({
+            "session_token": "tok_abc",
+            "session_id": "sid_123",
+            "scnt": "scnt_val"
+        });
+        std::fs::write(&path, data.to_string()).unwrap();
+
+        strip_session_routing_state(&path).await;
+
+        let contents = std::fs::read_to_string(&path).unwrap();
+        let map: HashMap<String, String> = serde_json::from_str(&contents).unwrap();
+        assert!(map.is_empty());
     }
 }
